AI-NLP/keyframe_extractor_classifier/module/Framify/requirements.txt
bson==0.5.10
ImageHash==4.2.1
numpy==1.19.5
#numpy==1.21.2
pandas==1.3.2
opencv-python==4.5.3.56
Pillow==8.3.2
scikit-image==0.18.3
scipy==1.7.1
